Refactor: MyPy 테스트 적용
Mypy 테스트 적용 후 통과를 위한 import 및 타입 처리
